+++ crates/deadpool/src/unmanaged/config.rs
@@ -35,6 +35,6 @@ impl Default for PoolConfig {
     /// Create a [`PoolConfig`] where [`PoolConfig::max_size`] is set to
     /// `cpu_count * 4` ignoring any logical CPUs (Hyper-Threading).
     fn default() -> Self {
-        Self::new(num_cpus::get_physical() * 4)
+        Self::new(crate::util::get_default_pool_max_size())
     }
 }